chore: #13 rename to snippets
