+++ convex/_generated/api.d.ts
@@ -15,7 +15,7 @@ import type {
 } from "convex/server";
 import type * as codeExecutions from "../codeExecutions.js";
 import type * as http from "../http.js";
-import type * as snippet from "../snippet.js";
+import type * as snippets from "../snippets.js";
 import type * as users from "../users.js";
 
 /**
@@ -29,7 +29,7 @@ import type * as users from "../users.js";
 declare const fullApi: ApiFromModules<{
   codeExecutions: typeof codeExecutions;
   http: typeof http;
-  snippet: typeof snippet;
+  snippets: typeof snippets;
   users: typeof users;
 }>;
 export declare const api: FilterApi<
