+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@clerk/clerk-react": "^5.21.0",
     "@clerk/nextjs": "^6.9.6",
     "@monaco-editor/react": "^4.6.0",
     "convex": "^1.17.4",
