+++ src/app/(root)/_components/editorPanel/EditorPanel.tsx
@@ -1,18 +1,19 @@
 "use client";
-
-import { useEffect, useState } from "react";
-import Image from "next/image";
 import { useCodeEditorStore } from "@/store/useCodeEditorStore";
-import { defineMonacoThemes, LANGUAGE_CONFIG } from "../../_constants";
-import { RotateCcwIcon, ShareIcon, TypeIcon } from "lucide-react";
+import { useEffect, useState } from "react";
+
 import { Editor } from "@monaco-editor/react";
 import { motion } from "framer-motion";
+import Image from "next/image";
+import { RotateCcwIcon, ShareIcon, TypeIcon } from "lucide-react";
 import { useClerk } from "@clerk/nextjs";
-import EditorPanelSkeleton from "../skeleton/editorPanelSkeleton/EditorPanelSkeleton";
-import ShareSnippetDialog from "../shareSnippetDialog/ShareSnippetDialog";
-import useMounted from "@/hooks/useMounted";
 
-const EditorPanel = () => {
+import useMounted from "@/hooks/useMounted";
+import ShareSnippetDialog from "../shareSnippetDialog/ShareSnippetDialog";
+import EditorPanelSkeleton from "../skeleton/editorPanelSkeleton/EditorPanelSkeleton";
+import { defineMonacoThemes, LANGUAGE_CONFIG } from "../../_constants";
+
+function EditorPanel() {
   const clerk = useClerk();
   const isMounted = useMounted();
   const [isShareDialogOpen, setIsShareDialogOpen] = useState(false);
@@ -21,43 +22,29 @@ const EditorPanel = () => {
 
   const handleRefresh = () => {
     const defaultCode = LANGUAGE_CONFIG[language].defaultCode;
-
-    if (editor) {
-      editor.setValue(defaultCode);
-    }
-
+    if (editor) editor.setValue(defaultCode);
     localStorage.removeItem(`editor-code-${language}`);
   };
 
-  const handleEditorChange = (value?: string) => {
-    if (value) {
-      localStorage.setItem(`editor-code-${language}`, value);
-    }
+  const handleEditorChange = (value: string | undefined) => {
+    if (value) localStorage.setItem(`editor-code-${language}`, value);
   };
 
   const handleFontSizeChange = (newSize: number) => {
-    // size => 12 - 24
     const size = Math.min(Math.max(newSize, 12), 24);
     setFontSize(size);
-
     localStorage.setItem("editor-font-size", size.toString());
   };
 
   useEffect(() => {
     const savedCode = localStorage.getItem(`editor-code-${language}`);
     const newCode = savedCode || LANGUAGE_CONFIG[language].defaultCode;
-
-    if (editor) {
-      setEditor(newCode);
-    }
-  }, [language, editor, setEditor]);
+    if (editor) editor.setValue(newCode);
+  }, [language, editor]);
 
   useEffect(() => {
     const savedFontSize = localStorage.getItem("editor-font-size");
-
-    if (savedFontSize) {
-      setFontSize(Number(savedFontSize));
-    }
+    if (savedFontSize) setFontSize(parseInt(savedFontSize));
   }, [setFontSize]);
 
   if (!isMounted) return null;
@@ -120,7 +107,7 @@ const EditorPanel = () => {
               whileTap={{ scale: 0.98 }}
               onClick={() => setIsShareDialogOpen(true)}
               className="inline-flex items-center gap-2 px-4 py-2 rounded-lg overflow-hidden bg-gradient-to-r
-           from-blue-500 to-blue-600 opacity-90 hover:opacity-100 transition-opacity"
+               from-blue-500 to-blue-600 opacity-90 hover:opacity-100 transition-opacity"
             >
               <ShareIcon className="size-4 text-white" />
               <span className="text-sm font-medium text-white ">Share</span>
@@ -170,6 +157,5 @@ const EditorPanel = () => {
       )}
     </div>
   );
-};
-
+}
 export default EditorPanel;
