+++ src/app/snippets/[id]/page.tsx
@@ -1,0 +1,5 @@
+const SnippetDetailPage = () => {
+  return <div></div>;
+};
+
+export default SnippetDetailPage;
